Moved model to its own folder
