--- __init__.py
@@ -1,5 +1,0 @@
-'''
-Created on 18 Jun 2018
-
-@author: jhkwakkel
-'''
