+++ model/__init__.py
@@ -1,0 +1,11 @@
+'''
+Created on 18 Jun 2018
+
+@author: jhkwakkel, jasonrwang
+
+This lets the model be run as a module with analysis elsewhere.
+'''
+
+import os
+import sys
+sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
