+++ model/funs_dikes.py
