+++ model/funs_generate_network.py
@@ -3,7 +3,7 @@ from __future__ import division, unicode_literals, print_function
 import numpy as np
 import networkx as nx
 import pandas as pd
-from funs_dikes import Lookuplin  # @UnresolvedImport
+from model.funs_dikes import Lookuplin  # @UnresolvedImport
 
 
 def to_dict_dropna(data):
